refactor(tui): extract dashboard tree renderers
